# Conflicts:
#	csc207/src/main/java/view/GoalView.java

+++ csc207/src/main/java/entity/Income.java
@@ -6,7 +6,7 @@ import java.util.Date;
  * Entity Income.
  */
 public class Income extends Transaction {
-    public Income(String name, int amount, String category, Date date) {
+    public Income(String name, double amount, String category, Date date) {
         super(name, amount, category, date);
     }
 }
